backend/requirements.txt
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
motor==3.7.0
pydantic==1.10.7
pymongo==4.11.3
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.1
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.13.0
uvicorn==0.21.1
